# target/generated-sources/openapi/src/main/java/com/autodebitapi/itau/api/CompareApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.11.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.autodebitapi.itau.api;

import com.autodebitapi.itau.api.model.CompareRequest;
import com.autodebitapi.itau.api.model.CompareResponse;
import com.autodebitapi.itau.api.model.ErrorResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Mono;

import jakarta.validation.Valid;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-24T01:21:42.654149-03:00[GMT-03:00]", comments = "Generator version: 7.11.0")
@Validated
@Tag(name = "Comparison", description = "Operações de comparação entre produtos")
public interface CompareApi {

    /**
     * POST /compare/products : Comparar produtos (Reativo)
     * Compara múltiplos produtos de forma assíncrona e retorna análise comparativa
     *
     * @param compareRequest  (required)
     * @return Comparação realizada com sucesso (status code 200)
     *         or Requisição inválida - dados de entrada incorretos (status code 400)
     *         or Recurso não encontrado (status code 404)
     *         or Erro interno do servidor (status code 500)
     */
    @Operation(
        operationId = "compareProducts",
        summary = "Comparar produtos (Reativo)",
        description = "Compara múltiplos produtos de forma assíncrona e retorna análise comparativa",
        tags = { "Comparison" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Comparação realizada com sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CompareResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requisição inválida - dados de entrada incorretos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Recurso não encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro interno do servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/compare/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CompareResponse>> _compareProducts(
        @Parameter(name = "CompareRequest", description = "", required = true) @Valid @RequestBody Mono<CompareRequest> compareRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return compareProducts(compareRequest, exchange);
    }

    // Override this method
    default  Mono<ResponseEntity<CompareResponse>> compareProducts(Mono<CompareRequest> compareRequest,  final ServerWebExchange exchange) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"comparison\" : { \"price\" : [ 1299.99, 1199.99, 699.99 ], \"rating\" : [ \"AWESOME\", \"AWESOME\", \"GOOD\" ], \"RAM\" : [ \"N/A\", \"N/A\", \"N/A\" ], \"Storage\" : [ \"128GB\", \"256GB\", \"128GB\" ] }, \"products\" : [ { \"price\" : 1299.99, \"name\" : \"iPhone 15 Pro\", \"rating\" : \"AWESOME\", \"description\" : \"Smartphone Apple com chip A17 Pro\", \"specification\" : \"128GB, Titânio Natural, 5G\", \"id\" : 1, \"url\" : \"https://apple.com/iphone15pro\" }, { \"price\" : 1299.99, \"name\" : \"iPhone 15 Pro\", \"rating\" : \"AWESOME\", \"description\" : \"Smartphone Apple com chip A17 Pro\", \"specification\" : \"128GB, Titânio Natural, 5G\", \"id\" : 1, \"url\" : \"https://apple.com/iphone15pro\" } ] }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(compareRequest).then(Mono.empty());

    }

}