# target/generated-sources/openapi/src/main/java/com/autodebitapi/itau/api/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.11.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.autodebitapi.itau.api;

import com.autodebitapi.itau.api.model.ErrorResponse;
import com.autodebitapi.itau.api.model.ProductRequest;
import com.autodebitapi.itau.api.model.ProductResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-24T01:21:42.654149-03:00[GMT-03:00]", comments = "Generator version: 7.11.0")
@Validated
@Tag(name = "Products", description = "Operações CRUD para gerenciamento de produtos")
public interface ProductsApi {

    /**
     * POST /products : Criar novo produto
     * Cria um novo produto no sistema
     *
     * @param productRequest  (required)
     * @return Produto criado com sucesso (status code 201)
     *         or Requisição inválida - dados de entrada incorretos (status code 400)
     *         or Erro interno do servidor (status code 500)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Criar novo produto",
        description = "Cria um novo produto no sistema",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Produto criado com sucesso"),
            @ApiResponse(responseCode = "400", description = "Requisição inválida - dados de entrada incorretos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro interno do servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<Void>> _createProduct(
        @Parameter(name = "ProductRequest", description = "", required = true) @Valid @RequestBody Mono<ProductRequest> productRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return createProduct(productRequest, exchange);
    }

    // Override this method
    default  Mono<ResponseEntity<Void>> createProduct(Mono<ProductRequest> productRequest,  final ServerWebExchange exchange) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(productRequest).then(Mono.empty());

    }


    /**
     * DELETE /products : Deletar todos os produtos
     * Remove todos os produtos do sistema
     *
     * @return Todos os produtos foram deletados com sucesso (status code 204)
     *         or Erro interno do servidor (status code 500)
     */
    @Operation(
        operationId = "deleteAllProducts",
        summary = "Deletar todos os produtos",
        description = "Remove todos os produtos do sistema",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Todos os produtos foram deletados com sucesso"),
            @ApiResponse(responseCode = "500", description = "Erro interno do servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/products",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<Void>> _deleteAllProducts(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return deleteAllProducts(exchange);
    }

    // Override this method
    default  Mono<ResponseEntity<Void>> deleteAllProducts( final ServerWebExchange exchange) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(Mono.empty());

    }


    /**
     * DELETE /products/{id} : Deletar produto por ID
     * Remove um produto específico do sistema
     *
     * @param id ID único do produto a ser deletado (required)
     * @return Produto deletado com sucesso (status code 204)
     *         or Recurso não encontrado (status code 404)
     *         or Erro interno do servidor (status code 500)
     */
    @Operation(
        operationId = "deleteProductById",
        summary = "Deletar produto por ID",
        description = "Remove um produto específico do sistema",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Produto deletado com sucesso"),
            @ApiResponse(responseCode = "404", description = "Recurso não encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro interno do servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/products/{id}",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<Void>> _deleteProductById(
        @Min(1L) @Parameter(name = "id", description = "ID único do produto a ser deletado", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return deleteProductById(id, exchange);
    }

    // Override this method
    default  Mono<ResponseEntity<Void>> deleteProductById(Long id,  final ServerWebExchange exchange) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(Mono.empty());

    }


    /**
     * GET /products : Listar todos os produtos
     * Retorna uma lista com todos os produtos cadastrados
     *
     * @return Lista de produtos retornada com sucesso (status code 200)
     *         or Erro interno do servidor (status code 500)
     */
    @Operation(
        operationId = "getAllProducts",
        summary = "Listar todos os produtos",
        description = "Retorna uma lista com todos os produtos cadastrados",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de produtos retornada com sucesso", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProductResponse.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Erro interno do servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<Flux<ProductResponse>>> _getAllProducts(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getAllProducts(exchange);
    }

    // Override this method
    default  Mono<ResponseEntity<Flux<ProductResponse>>> getAllProducts( final ServerWebExchange exchange) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "[ { \"price\" : 1299.99, \"name\" : \"iPhone 15 Pro\", \"rating\" : \"FIVE_STARS\", \"description\" : \"Smartphone Apple com chip A17 Pro e câmera de 48MP\", \"specification\" : \"128GB, Titânio Natural, 5G\", \"id\" : 1, \"url\" : \"https://apple.com/iphone15pro\" }, { \"price\" : 1299.99, \"name\" : \"iPhone 15 Pro\", \"rating\" : \"FIVE_STARS\", \"description\" : \"Smartphone Apple com chip A17 Pro e câmera de 48MP\", \"specification\" : \"128GB, Titânio Natural, 5G\", \"id\" : 1, \"url\" : \"https://apple.com/iphone15pro\" } ]";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(Mono.empty());

    }


    /**
     * GET /products/{id} : Buscar produto por ID
     * Retorna um produto específico pelo seu ID
     *
     * @param id ID único do produto (required)
     * @return Produto encontrado com sucesso (status code 200)
     *         or Recurso não encontrado (status code 404)
     *         or Erro interno do servidor (status code 500)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Buscar produto por ID",
        description = "Retorna um produto específico pelo seu ID",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Produto encontrado com sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Recurso não encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erro interno do servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/{id}",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<ProductResponse>> _getProductById(
        @Min(1L) @Parameter(name = "id", description = "ID único do produto", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getProductById(id, exchange);
    }

    // Override this method
    default  Mono<ResponseEntity<ProductResponse>> getProductById(Long id,  final ServerWebExchange exchange) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"price\" : 1299.99, \"name\" : \"iPhone 15 Pro\", \"rating\" : \"FIVE_STARS\", \"description\" : \"Smartphone Apple com chip A17 Pro e câmera de 48MP\", \"specification\" : \"128GB, Titânio Natural, 5G\", \"id\" : 1, \"url\" : \"https://apple.com/iphone15pro\" }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"message\" : \"Product not found\", \"status\" : 404 }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
        }
        return result.then(Mono.empty());

    }

}